requirements.txt
streamlit==1.51.0
pandas
numpy==2.3.3
altair==5.5.0
lxml==6.0.2
python-dotenv==1.2.1
openai==2.6.0
